% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Title}
\usage{
draw_names(the_name, the_name2)
}
\arguments{
\item{the_name}{name 1}

\item{the_name2}{name 2}
}
\value{
result
}
\description{
Title
}
\examples{
draw_names(the_name = 'Youssef', the_name2 = 'Margaux')
}
