% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_excel.R
\name{multi_excel}
\alias{multi_excel}
\title{Title}
\usage{
multi_excel(file)
}
\arguments{
\item{file}{This is the file from which we want to import the sheets}
}
\value{
a list
}
\description{
Title
}
\examples{
\dontrun{
multi_excel('jeux.xlsx')
}
}
