% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveascsv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Title}
\usage{
save_as_csv(dataset, filename, row.names = FALSE, ...)
}
\description{
Title
}
\examples{

data(iris)
save_as_csv(dataset=iris, filename="iris2.csv")

}
